+++ auth-db/queries/users/index.ts
@@ -1,3 +1,4 @@
 export * from "./create.ts";
 export * from "./get-all.ts";
 export * from "./get-by-email.ts";
+export * from "./get-by-id.ts";
